// src/lib/api/models/createVisibilityPolicyDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { CreateVisibilityPolicyDtoEntityType } from './createVisibilityPolicyDtoEntityType';
import type { VisibilityRuleDto } from './visibilityRuleDto';

export interface CreateVisibilityPolicyDto {
  name: string;
  entityType: CreateVisibilityPolicyDtoEntityType;
  priority?: number;
  rules?: VisibilityRuleDto[];
}