// src/lib/api/models/addVisibilityRuleDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { AddVisibilityRuleDtoVisibilityLevel } from './addVisibilityRuleDtoVisibilityLevel';
import type { AddVisibilityRuleDtoMaskType } from './addVisibilityRuleDtoMaskType';
import type { AddVisibilityRuleDtoConditionJson } from './addVisibilityRuleDtoConditionJson';

export interface AddVisibilityRuleDto {
  fieldKey: string;
  roleCode: string;
  visibilityLevel: AddVisibilityRuleDtoVisibilityLevel;
  maskType?: AddVisibilityRuleDtoMaskType;
  conditionJson?: AddVisibilityRuleDtoConditionJson;
}